Add test that loads CF standard name and units table and verifies against it
