+++ .github/workflows/code-quality.yml
@@ -5,7 +5,6 @@ permissions:
 
 on:
   pull_request:
-    branches: [ main ]
 
 jobs:
   quality:
